moved ecoverse list to separate component
