+++ src/pages/home/EcoverseOverview.tsx
@@ -1,18 +1,15 @@
 import React, { useMemo } from 'react';
 import { Col } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
-import { ReactComponent as Globe } from 'bootstrap-icons/icons/globe2.svg';
-import Section, { Body, Header as SectionHeader, SubHeader } from '../components/core/Section';
-import { CardContainer } from '../components/core/Container';
-import Icon from '../components/core/Icon';
-import Loading from '../components/core/Loading';
-import EcoverseCard from '../components/Ecoverse/EcoverseCard';
-import ErrorBlock from '../components/core/ErrorBlock';
-import { useUserContext } from '../hooks/useUserContext';
-import { useUpdateNavigation } from '../hooks/useNavigation';
-import { useEcoversesContext } from '../hooks/useEcoversesContext';
+import { CardContainer } from '../../components/core/Container';
+import Loading from '../../components/core/Loading';
+import EcoverseCard from '../../components/Ecoverse/EcoverseCard';
+import ErrorBlock from '../../components/core/ErrorBlock';
+import { useUserContext } from '../../hooks/useUserContext';
+import { useUpdateNavigation } from '../../hooks/useNavigation';
+import { useEcoversesContext } from '../../hooks/useEcoversesContext';
 
-const HomePage = () => {
+const EcoverseOverview = () => {
   const { t } = useTranslation();
   const { user } = useUserContext();
   const { ecoverses, loading, error } = useEcoversesContext();
@@ -20,18 +17,11 @@ const HomePage = () => {
   const currentPaths = useMemo(() => [], []);
   useUpdateNavigation({ currentPaths });
 
-  if (loading) {
-    return <Loading text={'Loading ecoverses'} />;
-  }
-
   return (
-    <>
-      <Section avatar={<Icon component={Globe} color="primary" size="xl" />}>
-        <SectionHeader text={t('pages.home.sections.ecoverse.header')} />
-        <SubHeader text={t('pages.home.sections.ecoverse.subheader')} />
-        <Body text={t('pages.home.sections.ecoverse.body')} />
-      </Section>
-      {error ? (
+    <React.Fragment>
+      {loading ? (
+        <Loading text={'Loading ecoverses'} />
+      ) : error ? (
         <Col xs={12}>
           <ErrorBlock blockName={t('pages.home.sections.ecoverse.header')} />
         </Col>
@@ -52,8 +42,8 @@ const HomePage = () => {
           ))}
         </CardContainer>
       )}
-    </>
+    </React.Fragment>
   );
 };
 
-export default HomePage;
+export default EcoverseOverview;
